requirements.txt
torch
torchvision
torchaudio
pytorch-lightning
hydra-core
omegaconf
wandb
timm
tqdm
requests
PyYAML
numpy
Pillow
opencv-python
datasets
pyarrow
imagecorruptions
matplotlib
git+https://github.com/facebookresearch/FFCV-SSL.git
